DataCollectornValidatorAgent/requirements.txt
# RAMESH Agent Requirements
# ==========================

# Core
openai>=1.0.0
python-dotenv

# Database & Memory
postgrest
numpy

# Z-Library Collection
playwright
fastmcp

# arXiv Collection (NEW!)
aiohttp>=3.8.0
aiofiles>=23.0.0

# Utils
python-levenshtein

# Install playwright browsers after pip install:
# playwright install chromium
